# OP-Card-Game-SIM/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: One Piece TCG Simulator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List open/active game rooms
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [games]
      summary: Create a new game room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Game room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get game room details
      responses:
        "200":
          description: Game details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGame
      tags: [games]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Close/delete a game room (host only, waiting status)
      responses:
        "204":
          description: Game deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /games/{id}/join:
    post:
      operationId: joinGame
      tags: [games]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Join an existing game room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinGameInput"
      responses:
        "200":
          description: Joined game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "400":
          description: Room full or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get current user stats (wins, losses, win rate, games played)
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserProfile:
      type: object
      required: [id, username, email, gamesPlayed, wins, losses, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        gamesPlayed:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        createdAt:
          type: string
          format: date-time
    UserStats:
      type: object
      required: [gamesPlayed, wins, losses, winRate]
      properties:
        gamesPlayed:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winRate:
          type: number
...
    UserProfileUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        avatarUrl:
          type: string
...
    Game:
      type: object
      required: [id, status, hostId, hostUsername, createdAt]
      properties:
        id:
          type: integer
        status:
          type: string
        isPrivate:
          type: boolean
        hostId:
          type: integer
        hostUsername:
          type: string
        hostDeckId:
          type: ["integer", "null"]
        guestId:
          type: ["integer", "null"]
        guestUsername:
          type: ["string", "null"]
        guestDeckId:
          type: ["integer", "null"]
        winnerId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: ["string", "null"]
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
    GameInput:
      type: object
      required: [deckId]
      properties:
        deckId:
          type: integer
        isPrivate:
          type: boolean
        isLocal:
          type: boolean
        guestDeckId:
          type: integer
    JoinGameInput:
      type: object
      required: [deckId]
      properties:
        deckId:
          type: integer