# OP-Card-Game-SIM/lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * One Piece TCG Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface UserStats {
  gamesPlayed: number;
  wins: number;
  losses: number;
  winRate: number;
}